Completed add a grade
